Map Food Consumption Score data variables to corresponding FCS food groups

Map Food Consumption Score data variables to corresponding FCS food groups

Usage

fcs_map_fg_vars(
  staples,
  pulses,
  vegetables,
  fruits,
  meat_fish,
  milk,
  sugar,
  oil,
  condiments
)

Arguments

staples, pulses, vegetables, fruits, meat_fish, milk, sugar, oil, condiments

A character value or vector of values for variables name/s for corresponding food groups.

Value

A named list of variable names for corresponding food groups.

Examples

vars <- paste0(
  "FCS", 
  c("Stap", "Pulse", "Veg", "Fruit", "Pr", "Dairy", "Sugar", "Fat", "Cond")
) |>
  as.list()

do.call(fcs_map_fg_vars, vars) 
#> $staples
#> [1] "FCSStap"
#> 
#> $pulses
#> [1] "FCSPulse"
#> 
#> $vegetables
#> [1] "FCSVeg"
#> 
#> $fruits
#> [1] "FCSFruit"
#> 
#> $meat_fish
#> [1] "FCSPr"
#> 
#> $milk
#> [1] "FCSDairy"
#> 
#> $sugar
#> [1] "FCSSugar"
#> 
#> $oil
#> [1] "FCSFat"
#> 
#> $condiments
#> [1] "FCSCond"
#>