Map data variables to corresponding indicators

Map data variables to corresponding indicators

Usage

map_variables(...)

fcs_fg_map_variables(
  ...,
  indicators = c("staples", "pulses", "vegetables", "fruits", "meatfish", "milk",
    "sugar", "oil", "condiment")
)

Arguments

...

Name-value pairs. Name gives the labels for indicators. The value should be the corresponding variable name in a dataset used for that indicator.

indicators

A character vector of indicator labels for a specific dietary intake indicator set.

Value

A named list of variable name/s for corresponding indicators.

Examples

## Variable names in fcs01 mapped to corresponding food group indicators
map_variables(
  staples = "FCSStap",
  pulses = "FCSPulse",
  vegetables = "FCSVeg",
  fruits = "FCSFruit",
  meatfish = "FCSPr",
  milk = "FCSDairy",
  sugar = "FCSSugar",
  oil = "FCSFat",
  condiment = "FCSCond"
)
#> $staples
#> [1] "FCSStap"
#> 
#> $pulses
#> [1] "FCSPulse"
#> 
#> $vegetables
#> [1] "FCSVeg"
#> 
#> $fruits
#> [1] "FCSFruit"
#> 
#> $meatfish
#> [1] "FCSPr"
#> 
#> $milk
#> [1] "FCSDairy"
#> 
#> $sugar
#> [1] "FCSSugar"
#> 
#> $oil
#> [1] "FCSFat"
#> 
#> $condiment
#> [1] "FCSCond"
#>